refactor(oidc): PR #22 レビュー指摘の反映
- SigningKeyStatus 型を lib/services/oidc/types.ts に分離。keys.ts と
  OidcClientsClient.tsx の型重複を解消（クライアントは jose 依存を踏まず
  types.ts からのみ import する）
- authorize.test.ts のモックに SigningKeyStatus 型を明示し、
  `as unknown as true` 等の discriminated union キャストを除去
